// src/app/Models/PersonalProposal/creditinformation.ts
export class Creditinformation {
    CLIENT_ID: number;
    ID: number;
    PROPOSAL_ID: number;
    IS_EXISTING_LOAN_WITH_SUB
    IS_EXISTING_LOAN_WITH_OTHER_BANKS
    IS_GUARANTOR_TO_LOAN_WITH_SUB
    IS_GUARANTOR_TO_LOAN_WITH_OTHER_BANK
    IS_EARLIAR_LOAN_HISTORY
    IS_ANY_DEPOSITES_WITH_SUB
    IS_LOAN_TAKEN_FOR_CLOSE_OTHER_LOANS
    BANK_NAME: string;
    OUTSTANDING_BANK_AMOUNT: number;
    IS_EARLIAR_LOAN_HISTORY_WITH_OTHER_BANK
    IS_ANY_DEPOSITES_WITH_OTHER_BANK
    TYPE: string
    APPLICANT_ID: number
    OTHER_LOAN_DETAILS = ""
    LOAN_GUARENTEE_COUNT: number
    LOAN_GUARENTEE_AMOUNT: number

    IS_NOC_OBTAINED_AND_ENCLOSED
}